always open and close connection

+++ src/main/java/dao/MemberDAO.java
@@ -9,20 +9,23 @@ import model.Member;
 
 public class MemberDAO {
 	
-	Connection connection = Database.getConnection();
+	Connection connection = null;
 	PreparedStatement preparedStatement = null;
 	
 	public Member getMemberByAthleteId(int athleteId)throws Exception {
 		Member member = null;
 		try {
-			preparedStatement = connection.prepareStatement("SELECT id, athleteId, accessToken FROM members where athleteId=?");
-			preparedStatement.setLong(1,athleteId);
-			ResultSet rs = preparedStatement.executeQuery();
-			if (rs.next()) {
-				member = new Member();
-				member.setId(rs.getInt("id"));
-				member.setAthleteId(rs.getInt("athleteId"));
-				member.setAccessToken(rs.getString("accessToken"));
+			connection = Database.getConnection();
+			if (connection != null) {
+				preparedStatement = connection.prepareStatement("SELECT id, athleteId, accessToken FROM members where athleteId=?");
+				preparedStatement.setLong(1,athleteId);
+				ResultSet rs = preparedStatement.executeQuery();
+				if (rs.next()) {
+					member = new Member();
+					member.setId(rs.getInt("id"));
+					member.setAthleteId(rs.getInt("athleteId"));
+					member.setAccessToken(rs.getString("accessToken"));
+				}
 			}
 		
 		} catch (Exception e) {
@@ -32,7 +35,7 @@ public class MemberDAO {
 			if (preparedStatement != null)
 				preparedStatement.close();
 			if (connection != null)
-			connection.close();
+				connection.close();
 		}
 		return member;
 	}
@@ -43,25 +46,28 @@ public class MemberDAO {
 			return false;
 
 		try {
-			Member memberDB = getMemberByAthleteId(member.getAthleteId());
-			if (memberDB != null) {	// update
-				String sql = "update members set accessToken=? where athleteId=?";
-				preparedStatement = connection.prepareStatement(sql);
-				preparedStatement.setLong(1, member.getAthleteId());
-				preparedStatement.setString(2, member.getAccessToken());
-	
-				int rowsAffected = preparedStatement.executeUpdate();
-				if (rowsAffected > 0)
-					return true;
-			}
-			else {	// insert
-				String sql = "insert into members (athleteId, accessToken) values (?,?)";
-				PreparedStatement preparedStatement = connection.prepareStatement(sql);
-				preparedStatement.setLong(1, member.getAthleteId());
-				preparedStatement.setString(2, member.getAccessToken());
-				int rowsAffected = preparedStatement.executeUpdate();
-				if (rowsAffected > 0)
-					return true;
+			connection = Database.getConnection();
+			if (connection != null) {
+				Member memberDB = getMemberByAthleteId(member.getAthleteId());
+				if (memberDB != null) {	// update
+					String sql = "update members set accessToken=? where athleteId=?";
+					preparedStatement = connection.prepareStatement(sql);
+					preparedStatement.setLong(1, member.getAthleteId());
+					preparedStatement.setString(2, member.getAccessToken());
+		
+					int rowsAffected = preparedStatement.executeUpdate();
+					if (rowsAffected > 0)
+						return true;
+				}
+				else {	// insert
+					String sql = "insert into members (athleteId, accessToken) values (?,?)";
+					PreparedStatement preparedStatement = connection.prepareStatement(sql);
+					preparedStatement.setLong(1, member.getAthleteId());
+					preparedStatement.setString(2, member.getAccessToken());
+					int rowsAffected = preparedStatement.executeUpdate();
+					if (rowsAffected > 0)
+						return true;
+				}
 			}
 		} catch (Exception e) {
 			throw e;
@@ -81,15 +87,18 @@ public class MemberDAO {
 			return false;
 		
 		try {
-			Member memberDB = getMemberByAthleteId(member.getAthleteId());
-			if (memberDB != null) {
-				String sql = "delete from members where athleteId=?";
-				preparedStatement = connection.prepareStatement(sql);
-				preparedStatement.setLong(1, member.getAthleteId());
-	
-				int rowsAffected = preparedStatement.executeUpdate();
-				if (rowsAffected > 0)
-					return true;
+			connection = Database.getConnection();
+			if (connection != null) {
+				Member memberDB = getMemberByAthleteId(member.getAthleteId());
+				if (memberDB != null) {
+					String sql = "delete from members where athleteId=?";
+					preparedStatement = connection.prepareStatement(sql);
+					preparedStatement.setLong(1, member.getAthleteId());
+		
+					int rowsAffected = preparedStatement.executeUpdate();
+					if (rowsAffected > 0)
+						return true;
+				}
 			}
 		} catch (Exception e) {
 			throw e;
@@ -98,7 +107,7 @@ public class MemberDAO {
 			if (preparedStatement != null)
 				preparedStatement.close();
 			if (connection != null)
-			connection.close();
+				connection.close();
 		}
 		
 		return false;
